requirements.txt
google-api-python-client==1.7.11
cassandra-driver==3.20.2
graphene==2.1.8
flask==1.1.1
flask-cors==3.0.8
flask-mail==0.9.1
requests
yfinance==0.1.54
lxml==4.4.2